Branch b1_6 b=14655 r=adilger, deen
Use __u64 instead of int for valid bits in mdc functions

+++ lustre/mdc/mdc_internal.h
@@ -20,7 +20,7 @@ void mdc_pack_req_body(struct ptlrpc_request *req, int offset,
 void mdc_pack_rep_body(struct ptlrpc_request *);
 void mdc_readdir_pack(struct ptlrpc_request *req, int offset, __u64 pg_off,
 		      __u32 size, struct ll_fid *mdc_fid);
-void mdc_getattr_pack(struct ptlrpc_request *req, int offset, int valid,
+void mdc_getattr_pack(struct ptlrpc_request *req, int offset, __u64 valid,
                       int flags, struct mdc_op_data *data);
 void mdc_setattr_pack(struct ptlrpc_request *req, int offset,
                       struct mdc_op_data *data,
@@ -43,7 +43,7 @@ void mdc_rename_pack(struct ptlrpc_request *req, int offset,
                      struct mdc_op_data *data,
                      const char *old, int oldlen, const char *new, int newlen);
 void mdc_close_pack(struct ptlrpc_request *req, int offset, struct obdo *oa,
-                    int valid, struct obd_client_handle *och);
+                    __u64 valid, struct obd_client_handle *och);
 void mdc_exit_request(struct client_obd *cli);
 void mdc_enter_request(struct client_obd *cli);
